Use TextLabelControlled for Diss Card level

+++ partyDarling/source/PartyDarling/include/Player.h
@@ -24,7 +24,7 @@ public:
 	float maxSpeed;
 
 	float experience;
-	int level;
+	float level;
 
 	PD_DissStats * dissStats;
 	bool wonLastDissBattle;
